UserProj/requirements.txt
Django==5.2
djangorestframework==3.16.1
confluent-kafka==2.13.0